fix(drive): version the compacted address-balance proof wire format
The bind-and-bound proof decoding change (#4165) rewrote what the
compacted address-balance-changes proof looks like on the wire — a single
GroveDB proof became the two-proof CompactedAddressBalanceProof bincode
envelope — while leaving the method version registry at 0 on both the
prove and verify sides. With the format mutated in place there was no way
to distinguish old-format from new-format peers at runtime: any
client/server pair not built from the same commit failed this query by
construction (4.1 clients against 4.0-rc nodes and vice versa).

Restore the registry discipline by making the format protocol-version
negotiated:

- verify: v0 is the legacy single-proof decoder again (restored
  byte-for-byte wire-compatible, with the decode-allocation bounds from
  #4165 applied since they do not change the accepted format); the
  envelope decoder moves to v1. Registry: new
  DRIVE_VERIFY_METHOD_VERSIONS_V2 bumps
  verify_compacted_address_balance_changes to 1, wired only into
  DRIVE_VERSION_V8 (protocol version 13).
- prove: v0 emits the legacy single proof again; the envelope encoder
  moves to v1, dispatched on a new dedicated
  prove_compacted_address_balance_changes feature version (0 in drive
  bundles v1-v7, 1 in v8) so the fetch path's version is no longer
  overloaded for the proof format.
- Since drive-abci serves queries with the network's active epoch
  platform version and SDK clients track that same version via the proven
  metadata ratchet, both sides now switch formats together when protocol
  version 13 activates; mixed 4.0/4.1 networks interoperate on the legacy
  format until then.
- The predecessor-binding soundness fix inherently requires the envelope,
  so it remains v1-only; v0 keeps the legacy (proof-derived start key)
  semantics its deployed peers expect, documented as such.

Tests: legacy roundtrip pinned under protocol version 12, envelope
roundtrip under 13, and cross-format rejection in both directions
(envelope proof under a legacy verifier, legacy proof under the envelope
verifier).
