+++ packages/rs-platform-version/src/version/drive_versions/drive_verify_method_versions/mod.rs
@@ -1,6 +1,7 @@
 use versioned_feature_core::FeatureVersion;
 
 pub mod v1;
+pub mod v2;
 
 #[derive(Clone, Debug, Default)]
 pub struct DriveVerifyMethodVersions {
